refactor: break import cycles with shared type modules
