+++ src/components/customizer/tabs/tabTypes.ts
@@ -1,0 +1,5 @@
+export interface TabDisplayInfo {
+  readonly id: string;
+  readonly name: string;
+  readonly isModified?: boolean;
+}
