+++ src/types/campaign/skills/index.ts
@@ -8,5 +8,5 @@ export * from './IAttributes';
 export * from './ISkillType';
 export * from './ISkill';
 export * from './experienceLevels';
-export * from '../../../constants/campaign/skillCatalog';
+export * from '@/constants/campaign/skillCatalog';
 export * from './techSkill';
