+++ src/types/simulation-viewer/index.ts
@@ -4,3 +4,4 @@ export * from './ICampaignDashboardMetrics';
 export * from './IBattleComparisonData';
 export * from './IAnomalyThresholds';
 export * from './IDashboardConfig';
+export * from './IViolation';
